% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared-funs.R
\name{Ceig}
\alias{Ceig}
\title{Calculate Eigenvalues and Eigenvectors of Symmetric Matrix}
\usage{
Ceig(x, number = 1, tol = .Machine$double.eps)
}
\arguments{
\item{x}{Symmetric square matrix.}

\item{number}{Integer, the number of eigenvalues and vectors to compute.}

\item{tol}{Numeric, tolerance for determining positive values.}
}
\value{
A list with eigenvalues (\code{values}) and eigenvectors (\code{vectors}).
}
\description{
Computes the eigenvalues and eigenvectors of a symmetric matrix using its lower triangle.
Optimized by a C interface for efficiency.
}
