% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared-funs.R
\name{dist.colspace}
\alias{dist.colspace}
\title{Distance Between Column Spaces}
\usage{
dist.colspace(matA, matB)
}
\arguments{
\item{matA}{First orthonormal matrix.}

\item{matB}{Second orthonormal matrix.}
}
\value{
Distance between the column spaces of matA and matB.
}
\description{
Calculates the distance between the column spaces of two orthonormal matrices.
}
