% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foldedSIR.R
\name{foldedSIR}
\alias{foldedSIR}
\title{Folded Sliced Inverse Regression (SIR)}
\usage{
foldedSIR(y, x, p, nslice, d, maxiter = 200, tol = 1e-08)
}
\arguments{
\item{y}{Numeric vector of response variables, with one value per observation.}

\item{x}{Numeric matrix of predictor variables, with each row representing an observation.}

\item{p}{Numeric vector of length 2 specifying the dimensions into which \code{x} can be folded.}

\item{nslice}{Integer specifying the number of slices to divide \code{y} into for SIR analysis.}

\item{d}{Numeric vector of length 2 specifying the dimensions for dimensionality reduction.}

\item{maxiter}{Maximum number of iterations for the algorithm (default = 200).}

\item{tol}{Tolerance level for algorithm convergence (default = 1e-8).}
}
\value{
A list containing:
\itemize{
\item Various components related to the fitted model, including the number of iterations (\code{iter}) and convergence status (\code{conv}).
}
}
\description{
Implements a folded version of Sliced Inverse Regression (SIR) for dimensionality reduction,
designed for bidimensional tensorial data. This function requires the data matrix \code{x} to
have dimensions that can be folded into a tensor of shape specified by \code{p}.
}
\examples{
set.seed(123)
y <- rnorm(100)
x <- matrix(rnorm(400), ncol = 4)
p <- c(2, 2)
nslice <- 10
d <- c(1, 1)

result <- foldedSIR(y, x, p, nslice, d)
print(result)

}
