% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locpoly.R, R/shared-funs.R, R/testing.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{dr.slices}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dr}{\code{\link[dr]{dr.slices}}, \code{\link[dr]{dr.slices}}, \code{\link[dr]{dr.slices}}}
}}

