% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TensorSIR.R
\name{tensor2SIR}
\alias{tensor2SIR}
\title{Tensor-based SIR for Matrix-valued Predictors Using Naive Algorithm}
\usage{
tensor2SIR(prob)
}
\arguments{
\item{prob}{A list with problem-specific parameters and moment matrices.}
}
\value{
A list containing the tensor-based directions (\code{direction}), their Q-values (\code{Qvalue}),
and matrices \code{a} and \code{b} representing the tensor decomposition.
}
\description{
Extends tensor-based Sliced Inverse Regression (SIR) to handle matrix-valued predictors
by performing the analysis in a tensor framework, specifically for two dimensions (\code{p} = 2).
}
\examples{
# For usage, see the example for `foldedSIR`.

}
