% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{wchisq.tail}
\alias{wchisq.tail}
\title{Evaluate P-Value for Weighted Chi-Squared Distribution}
\usage{
wchisq.tail(stat, zvec, n.mc = 10000, eps = 1e-07)
}
\arguments{
\item{stat}{Numeric, the test statistic for which to evaluate the tail probability.}

\item{zvec}{Numeric vector of eigenvalues used in the weighting.}

\item{n.mc}{Integer, the number of Monte Carlo simulations to perform (default = 10000).}

\item{eps}{Numeric, a small tolerance value to filter \code{zvec} elements (default = 1e-7).}
}
\value{
Numeric, the estimated tail probability.
}
\description{
Calculates the tail probability of a weighted chi-squared distribution,
given a test statistic, eigenvalues vector \code{zvec}, and the number of Monte Carlo simulations.
}
